+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# x.y.z
+
+- Add customization props `previewForm` & `reviewForm`
+
 # 0.4.108
 
 - Remove confusing "Try again later" error message